fix: rename model cards and implementations consistently
