+++ discojs/src/models/cards/MNISTClassifier.ts
@@ -1,7 +1,7 @@
 import { Model, ModelCard } from "../index.js";
-import { model } from "../implementations/pytorch_mnist_classifier.js";
+import { model } from "../implementations/MNISTClassifierModel.js";
 
-export const PytorchMNISTClassifier: ModelCard<"image"> = {
+export const MNISTClassifier: ModelCard<"image"> = {
   card: {
     id: "mnist_classifier",
     name: "MNIST Classifier",
