+++ discojs/src/models/cards/index.ts
@@ -1,7 +1,6 @@
 export { LUSClassifier } from "./LUSClassifier.js";
-export { CIFAR10Classifier } from "./mobileNetV1_cifar10.js";
-export { PytorchMNISTClassifier } from "./pytorchMnistClassifier.js";
-export { FaceClassifier } from "./faceClassifier.js";
+export { CIFAR10Classifier } from "./CIFAR10Classifier.js";
+export { MNISTClassifier } from "./MNISTClassifier.js";
 export { DogClassifier } from "./dogClassifier.js";
 export { TitanicClassifier } from "./titanicClassifier.js";
 export { Wikitext } from "./wikitext.js";
