+++ discojs/src/models/implementations/CIFAR10ClassifierModel.ts
