+++ discojs/src/models/implementations/MNISTClassifierModel.ts
