+++ discojs/src/models/implementations/dogClassifierModel.ts
