--- discojs/src/models/implementations/mobileNetV2_35_alpha_2_classes.ts
@@ -1,4303 +1,0 @@
-// Source: https://storage.googleapis.com/deai-313515.appspot.com/models/mobileNetV2_35_alpha_2_classes/model.json
-// This model was converted using the tensorflow.js converter
-export default {
-  format: "layers-model",
-  generatedBy: "keras v2.6.0",
-  convertedBy: "TensorFlow.js Converter v3.18.0",
-  modelTopology: {
-    keras_version: "2.6.0",
-    backend: "tensorflow",
-    model_config: {
-      class_name: "Functional",
-      config: {
-        name: "model",
-        layers: [
-          {
-            class_name: "InputLayer",
-            config: {
-              batch_input_shape: [null, null, null, 3],
-              dtype: "float32",
-              sparse: false,
-              ragged: false,
-              name: "input_1",
-            },
-            name: "input_1",
-            inbound_nodes: [],
-          },
-          {
-            class_name: "Conv2D",
-            config: {
-              name: "Conv1",
-              trainable: true,
-              dtype: "float32",
-              filters: 16,
-              kernel_size: [3, 3],
-              strides: [2, 2],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              kernel_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              bias_initializer: { class_name: "Zeros", config: {} },
-              kernel_regularizer: null,
-              bias_regularizer: null,
-              activity_regularizer: null,
-              kernel_constraint: null,
-              bias_constraint: null,
-            },
-            name: "Conv1",
-            inbound_nodes: [[["input_1", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "bn_Conv1",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "bn_Conv1",
-            inbound_nodes: [[["Conv1", 0, 0, {}]]],
-          },
-          {
-            class_name: "ReLU",
-            config: {
-              name: "Conv1_relu",
-              trainable: true,
-              dtype: "float32",
-              max_value: 6.0,
-              negative_slope: 0.0,
-              threshold: 0.0,
-            },
-            name: "Conv1_relu",
-            inbound_nodes: [[["bn_Conv1", 0, 0, {}]]],
-          },
-          {
-            class_name: "DepthwiseConv2D",
-            config: {
-              name: "expanded_conv_depthwise",
-              trainable: true,
-              dtype: "float32",
-              kernel_size: [3, 3],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              bias_initializer: { class_name: "Zeros", config: {} },
-              bias_regularizer: null,
-              activity_regularizer: null,
-              bias_constraint: null,
-              depth_multiplier: 1,
-              depthwise_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              depthwise_regularizer: null,
-              depthwise_constraint: null,
-            },
-            name: "expanded_conv_depthwise",
-            inbound_nodes: [[["Conv1_relu", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "expanded_conv_depthwise_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "expanded_conv_depthwise_BN",
-            inbound_nodes: [[["expanded_conv_depthwise", 0, 0, {}]]],
-          },
-          {
-            class_name: "ReLU",
-            config: {
-              name: "expanded_conv_depthwise_relu",
-              trainable: true,
-              dtype: "float32",
-              max_value: 6.0,
-              negative_slope: 0.0,
-              threshold: 0.0,
-            },
-            name: "expanded_conv_depthwise_relu",
-            inbound_nodes: [[["expanded_conv_depthwise_BN", 0, 0, {}]]],
-          },
-          {
-            class_name: "Conv2D",
-            config: {
-              name: "expanded_conv_project",
-              trainable: true,
-              dtype: "float32",
-              filters: 8,
-              kernel_size: [1, 1],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              kernel_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              bias_initializer: { class_name: "Zeros", config: {} },
-              kernel_regularizer: null,
-              bias_regularizer: null,
-              activity_regularizer: null,
-              kernel_constraint: null,
-              bias_constraint: null,
-            },
-            name: "expanded_conv_project",
-            inbound_nodes: [[["expanded_conv_depthwise_relu", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "expanded_conv_project_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "expanded_conv_project_BN",
-            inbound_nodes: [[["expanded_conv_project", 0, 0, {}]]],
-          },
-          {
-            class_name: "Conv2D",
-            config: {
-              name: "block_1_expand",
-              trainable: true,
-              dtype: "float32",
-              filters: 48,
-              kernel_size: [1, 1],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              kernel_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              bias_initializer: { class_name: "Zeros", config: {} },
-              kernel_regularizer: null,
-              bias_regularizer: null,
-              activity_regularizer: null,
-              kernel_constraint: null,
-              bias_constraint: null,
-            },
-            name: "block_1_expand",
-            inbound_nodes: [[["expanded_conv_project_BN", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_1_expand_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_1_expand_BN",
-            inbound_nodes: [[["block_1_expand", 0, 0, {}]]],
-          },
-          {
-            class_name: "ReLU",
-            config: {
-              name: "block_1_expand_relu",
-              trainable: true,
-              dtype: "float32",
-              max_value: 6.0,
-              negative_slope: 0.0,
-              threshold: 0.0,
-            },
-            name: "block_1_expand_relu",
-            inbound_nodes: [[["block_1_expand_BN", 0, 0, {}]]],
-          },
-          {
-            class_name: "ZeroPadding2D",
-            config: {
-              name: "block_1_pad",
-              trainable: true,
-              dtype: "float32",
-              padding: [
-                [0, 1],
-                [0, 1],
-              ],
-              data_format: "channels_last",
-            },
-            name: "block_1_pad",
-            inbound_nodes: [[["block_1_expand_relu", 0, 0, {}]]],
-          },
-          {
-            class_name: "DepthwiseConv2D",
-            config: {
-              name: "block_1_depthwise",
-              trainable: true,
-              dtype: "float32",
-              kernel_size: [3, 3],
-              strides: [2, 2],
-              padding: "valid",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              bias_initializer: { class_name: "Zeros", config: {} },
-              bias_regularizer: null,
-              activity_regularizer: null,
-              bias_constraint: null,
-              depth_multiplier: 1,
-              depthwise_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              depthwise_regularizer: null,
-              depthwise_constraint: null,
-            },
-            name: "block_1_depthwise",
-            inbound_nodes: [[["block_1_pad", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_1_depthwise_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_1_depthwise_BN",
-            inbound_nodes: [[["block_1_depthwise", 0, 0, {}]]],
-          },
-          {
-            class_name: "ReLU",
-            config: {
-              name: "block_1_depthwise_relu",
-              trainable: true,
-              dtype: "float32",
-              max_value: 6.0,
-              negative_slope: 0.0,
-              threshold: 0.0,
-            },
-            name: "block_1_depthwise_relu",
-            inbound_nodes: [[["block_1_depthwise_BN", 0, 0, {}]]],
-          },
-          {
-            class_name: "Conv2D",
-            config: {
-              name: "block_1_project",
-              trainable: true,
-              dtype: "float32",
-              filters: 8,
-              kernel_size: [1, 1],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              kernel_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              bias_initializer: { class_name: "Zeros", config: {} },
-              kernel_regularizer: null,
-              bias_regularizer: null,
-              activity_regularizer: null,
-              kernel_constraint: null,
-              bias_constraint: null,
-            },
-            name: "block_1_project",
-            inbound_nodes: [[["block_1_depthwise_relu", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_1_project_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_1_project_BN",
-            inbound_nodes: [[["block_1_project", 0, 0, {}]]],
-          },
-          {
-            class_name: "Conv2D",
-            config: {
-              name: "block_2_expand",
-              trainable: true,
-              dtype: "float32",
-              filters: 48,
-              kernel_size: [1, 1],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              kernel_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              bias_initializer: { class_name: "Zeros", config: {} },
-              kernel_regularizer: null,
-              bias_regularizer: null,
-              activity_regularizer: null,
-              kernel_constraint: null,
-              bias_constraint: null,
-            },
-            name: "block_2_expand",
-            inbound_nodes: [[["block_1_project_BN", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_2_expand_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_2_expand_BN",
-            inbound_nodes: [[["block_2_expand", 0, 0, {}]]],
-          },
-          {
-            class_name: "ReLU",
-            config: {
-              name: "block_2_expand_relu",
-              trainable: true,
-              dtype: "float32",
-              max_value: 6.0,
-              negative_slope: 0.0,
-              threshold: 0.0,
-            },
-            name: "block_2_expand_relu",
-            inbound_nodes: [[["block_2_expand_BN", 0, 0, {}]]],
-          },
-          {
-            class_name: "DepthwiseConv2D",
-            config: {
-              name: "block_2_depthwise",
-              trainable: true,
-              dtype: "float32",
-              kernel_size: [3, 3],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              bias_initializer: { class_name: "Zeros", config: {} },
-              bias_regularizer: null,
-              activity_regularizer: null,
-              bias_constraint: null,
-              depth_multiplier: 1,
-              depthwise_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              depthwise_regularizer: null,
-              depthwise_constraint: null,
-            },
-            name: "block_2_depthwise",
-            inbound_nodes: [[["block_2_expand_relu", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_2_depthwise_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_2_depthwise_BN",
-            inbound_nodes: [[["block_2_depthwise", 0, 0, {}]]],
-          },
-          {
-            class_name: "ReLU",
-            config: {
-              name: "block_2_depthwise_relu",
-              trainable: true,
-              dtype: "float32",
-              max_value: 6.0,
-              negative_slope: 0.0,
-              threshold: 0.0,
-            },
-            name: "block_2_depthwise_relu",
-            inbound_nodes: [[["block_2_depthwise_BN", 0, 0, {}]]],
-          },
-          {
-            class_name: "Conv2D",
-            config: {
-              name: "block_2_project",
-              trainable: true,
-              dtype: "float32",
-              filters: 8,
-              kernel_size: [1, 1],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              kernel_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              bias_initializer: { class_name: "Zeros", config: {} },
-              kernel_regularizer: null,
-              bias_regularizer: null,
-              activity_regularizer: null,
-              kernel_constraint: null,
-              bias_constraint: null,
-            },
-            name: "block_2_project",
-            inbound_nodes: [[["block_2_depthwise_relu", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_2_project_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_2_project_BN",
-            inbound_nodes: [[["block_2_project", 0, 0, {}]]],
-          },
-          {
-            class_name: "Add",
-            config: { name: "block_2_add", trainable: true, dtype: "float32" },
-            name: "block_2_add",
-            inbound_nodes: [
-              [
-                ["block_1_project_BN", 0, 0, {}],
-                ["block_2_project_BN", 0, 0, {}],
-              ],
-            ],
-          },
-          {
-            class_name: "Conv2D",
-            config: {
-              name: "block_3_expand",
-              trainable: true,
-              dtype: "float32",
-              filters: 48,
-              kernel_size: [1, 1],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              kernel_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              bias_initializer: { class_name: "Zeros", config: {} },
-              kernel_regularizer: null,
-              bias_regularizer: null,
-              activity_regularizer: null,
-              kernel_constraint: null,
-              bias_constraint: null,
-            },
-            name: "block_3_expand",
-            inbound_nodes: [[["block_2_add", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_3_expand_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_3_expand_BN",
-            inbound_nodes: [[["block_3_expand", 0, 0, {}]]],
-          },
-          {
-            class_name: "ReLU",
-            config: {
-              name: "block_3_expand_relu",
-              trainable: true,
-              dtype: "float32",
-              max_value: 6.0,
-              negative_slope: 0.0,
-              threshold: 0.0,
-            },
-            name: "block_3_expand_relu",
-            inbound_nodes: [[["block_3_expand_BN", 0, 0, {}]]],
-          },
-          {
-            class_name: "ZeroPadding2D",
-            config: {
-              name: "block_3_pad",
-              trainable: true,
-              dtype: "float32",
-              padding: [
-                [0, 1],
-                [0, 1],
-              ],
-              data_format: "channels_last",
-            },
-            name: "block_3_pad",
-            inbound_nodes: [[["block_3_expand_relu", 0, 0, {}]]],
-          },
-          {
-            class_name: "DepthwiseConv2D",
-            config: {
-              name: "block_3_depthwise",
-              trainable: true,
-              dtype: "float32",
-              kernel_size: [3, 3],
-              strides: [2, 2],
-              padding: "valid",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              bias_initializer: { class_name: "Zeros", config: {} },
-              bias_regularizer: null,
-              activity_regularizer: null,
-              bias_constraint: null,
-              depth_multiplier: 1,
-              depthwise_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              depthwise_regularizer: null,
-              depthwise_constraint: null,
-            },
-            name: "block_3_depthwise",
-            inbound_nodes: [[["block_3_pad", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_3_depthwise_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_3_depthwise_BN",
-            inbound_nodes: [[["block_3_depthwise", 0, 0, {}]]],
-          },
-          {
-            class_name: "ReLU",
-            config: {
-              name: "block_3_depthwise_relu",
-              trainable: true,
-              dtype: "float32",
-              max_value: 6.0,
-              negative_slope: 0.0,
-              threshold: 0.0,
-            },
-            name: "block_3_depthwise_relu",
-            inbound_nodes: [[["block_3_depthwise_BN", 0, 0, {}]]],
-          },
-          {
-            class_name: "Conv2D",
-            config: {
-              name: "block_3_project",
-              trainable: true,
-              dtype: "float32",
-              filters: 16,
-              kernel_size: [1, 1],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              kernel_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              bias_initializer: { class_name: "Zeros", config: {} },
-              kernel_regularizer: null,
-              bias_regularizer: null,
-              activity_regularizer: null,
-              kernel_constraint: null,
-              bias_constraint: null,
-            },
-            name: "block_3_project",
-            inbound_nodes: [[["block_3_depthwise_relu", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_3_project_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_3_project_BN",
-            inbound_nodes: [[["block_3_project", 0, 0, {}]]],
-          },
-          {
-            class_name: "Conv2D",
-            config: {
-              name: "block_4_expand",
-              trainable: true,
-              dtype: "float32",
-              filters: 96,
-              kernel_size: [1, 1],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              kernel_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              bias_initializer: { class_name: "Zeros", config: {} },
-              kernel_regularizer: null,
-              bias_regularizer: null,
-              activity_regularizer: null,
-              kernel_constraint: null,
-              bias_constraint: null,
-            },
-            name: "block_4_expand",
-            inbound_nodes: [[["block_3_project_BN", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_4_expand_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_4_expand_BN",
-            inbound_nodes: [[["block_4_expand", 0, 0, {}]]],
-          },
-          {
-            class_name: "ReLU",
-            config: {
-              name: "block_4_expand_relu",
-              trainable: true,
-              dtype: "float32",
-              max_value: 6.0,
-              negative_slope: 0.0,
-              threshold: 0.0,
-            },
-            name: "block_4_expand_relu",
-            inbound_nodes: [[["block_4_expand_BN", 0, 0, {}]]],
-          },
-          {
-            class_name: "DepthwiseConv2D",
-            config: {
-              name: "block_4_depthwise",
-              trainable: true,
-              dtype: "float32",
-              kernel_size: [3, 3],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              bias_initializer: { class_name: "Zeros", config: {} },
-              bias_regularizer: null,
-              activity_regularizer: null,
-              bias_constraint: null,
-              depth_multiplier: 1,
-              depthwise_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              depthwise_regularizer: null,
-              depthwise_constraint: null,
-            },
-            name: "block_4_depthwise",
-            inbound_nodes: [[["block_4_expand_relu", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_4_depthwise_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_4_depthwise_BN",
-            inbound_nodes: [[["block_4_depthwise", 0, 0, {}]]],
-          },
-          {
-            class_name: "ReLU",
-            config: {
-              name: "block_4_depthwise_relu",
-              trainable: true,
-              dtype: "float32",
-              max_value: 6.0,
-              negative_slope: 0.0,
-              threshold: 0.0,
-            },
-            name: "block_4_depthwise_relu",
-            inbound_nodes: [[["block_4_depthwise_BN", 0, 0, {}]]],
-          },
-          {
-            class_name: "Conv2D",
-            config: {
-              name: "block_4_project",
-              trainable: true,
-              dtype: "float32",
-              filters: 16,
-              kernel_size: [1, 1],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              kernel_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              bias_initializer: { class_name: "Zeros", config: {} },
-              kernel_regularizer: null,
-              bias_regularizer: null,
-              activity_regularizer: null,
-              kernel_constraint: null,
-              bias_constraint: null,
-            },
-            name: "block_4_project",
-            inbound_nodes: [[["block_4_depthwise_relu", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_4_project_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_4_project_BN",
-            inbound_nodes: [[["block_4_project", 0, 0, {}]]],
-          },
-          {
-            class_name: "Add",
-            config: { name: "block_4_add", trainable: true, dtype: "float32" },
-            name: "block_4_add",
-            inbound_nodes: [
-              [
-                ["block_3_project_BN", 0, 0, {}],
-                ["block_4_project_BN", 0, 0, {}],
-              ],
-            ],
-          },
-          {
-            class_name: "Conv2D",
-            config: {
-              name: "block_5_expand",
-              trainable: true,
-              dtype: "float32",
-              filters: 96,
-              kernel_size: [1, 1],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              kernel_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              bias_initializer: { class_name: "Zeros", config: {} },
-              kernel_regularizer: null,
-              bias_regularizer: null,
-              activity_regularizer: null,
-              kernel_constraint: null,
-              bias_constraint: null,
-            },
-            name: "block_5_expand",
-            inbound_nodes: [[["block_4_add", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_5_expand_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_5_expand_BN",
-            inbound_nodes: [[["block_5_expand", 0, 0, {}]]],
-          },
-          {
-            class_name: "ReLU",
-            config: {
-              name: "block_5_expand_relu",
-              trainable: true,
-              dtype: "float32",
-              max_value: 6.0,
-              negative_slope: 0.0,
-              threshold: 0.0,
-            },
-            name: "block_5_expand_relu",
-            inbound_nodes: [[["block_5_expand_BN", 0, 0, {}]]],
-          },
-          {
-            class_name: "DepthwiseConv2D",
-            config: {
-              name: "block_5_depthwise",
-              trainable: true,
-              dtype: "float32",
-              kernel_size: [3, 3],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              bias_initializer: { class_name: "Zeros", config: {} },
-              bias_regularizer: null,
-              activity_regularizer: null,
-              bias_constraint: null,
-              depth_multiplier: 1,
-              depthwise_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              depthwise_regularizer: null,
-              depthwise_constraint: null,
-            },
-            name: "block_5_depthwise",
-            inbound_nodes: [[["block_5_expand_relu", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_5_depthwise_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_5_depthwise_BN",
-            inbound_nodes: [[["block_5_depthwise", 0, 0, {}]]],
-          },
-          {
-            class_name: "ReLU",
-            config: {
-              name: "block_5_depthwise_relu",
-              trainable: true,
-              dtype: "float32",
-              max_value: 6.0,
-              negative_slope: 0.0,
-              threshold: 0.0,
-            },
-            name: "block_5_depthwise_relu",
-            inbound_nodes: [[["block_5_depthwise_BN", 0, 0, {}]]],
-          },
-          {
-            class_name: "Conv2D",
-            config: {
-              name: "block_5_project",
-              trainable: true,
-              dtype: "float32",
-              filters: 16,
-              kernel_size: [1, 1],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              kernel_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              bias_initializer: { class_name: "Zeros", config: {} },
-              kernel_regularizer: null,
-              bias_regularizer: null,
-              activity_regularizer: null,
-              kernel_constraint: null,
-              bias_constraint: null,
-            },
-            name: "block_5_project",
-            inbound_nodes: [[["block_5_depthwise_relu", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_5_project_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_5_project_BN",
-            inbound_nodes: [[["block_5_project", 0, 0, {}]]],
-          },
-          {
-            class_name: "Add",
-            config: { name: "block_5_add", trainable: true, dtype: "float32" },
-            name: "block_5_add",
-            inbound_nodes: [
-              [
-                ["block_4_add", 0, 0, {}],
-                ["block_5_project_BN", 0, 0, {}],
-              ],
-            ],
-          },
-          {
-            class_name: "Conv2D",
-            config: {
-              name: "block_6_expand",
-              trainable: true,
-              dtype: "float32",
-              filters: 96,
-              kernel_size: [1, 1],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              kernel_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              bias_initializer: { class_name: "Zeros", config: {} },
-              kernel_regularizer: null,
-              bias_regularizer: null,
-              activity_regularizer: null,
-              kernel_constraint: null,
-              bias_constraint: null,
-            },
-            name: "block_6_expand",
-            inbound_nodes: [[["block_5_add", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_6_expand_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_6_expand_BN",
-            inbound_nodes: [[["block_6_expand", 0, 0, {}]]],
-          },
-          {
-            class_name: "ReLU",
-            config: {
-              name: "block_6_expand_relu",
-              trainable: true,
-              dtype: "float32",
-              max_value: 6.0,
-              negative_slope: 0.0,
-              threshold: 0.0,
-            },
-            name: "block_6_expand_relu",
-            inbound_nodes: [[["block_6_expand_BN", 0, 0, {}]]],
-          },
-          {
-            class_name: "ZeroPadding2D",
-            config: {
-              name: "block_6_pad",
-              trainable: true,
-              dtype: "float32",
-              padding: [
-                [0, 1],
-                [0, 1],
-              ],
-              data_format: "channels_last",
-            },
-            name: "block_6_pad",
-            inbound_nodes: [[["block_6_expand_relu", 0, 0, {}]]],
-          },
-          {
-            class_name: "DepthwiseConv2D",
-            config: {
-              name: "block_6_depthwise",
-              trainable: true,
-              dtype: "float32",
-              kernel_size: [3, 3],
-              strides: [2, 2],
-              padding: "valid",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              bias_initializer: { class_name: "Zeros", config: {} },
-              bias_regularizer: null,
-              activity_regularizer: null,
-              bias_constraint: null,
-              depth_multiplier: 1,
-              depthwise_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              depthwise_regularizer: null,
-              depthwise_constraint: null,
-            },
-            name: "block_6_depthwise",
-            inbound_nodes: [[["block_6_pad", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_6_depthwise_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_6_depthwise_BN",
-            inbound_nodes: [[["block_6_depthwise", 0, 0, {}]]],
-          },
-          {
-            class_name: "ReLU",
-            config: {
-              name: "block_6_depthwise_relu",
-              trainable: true,
-              dtype: "float32",
-              max_value: 6.0,
-              negative_slope: 0.0,
-              threshold: 0.0,
-            },
-            name: "block_6_depthwise_relu",
-            inbound_nodes: [[["block_6_depthwise_BN", 0, 0, {}]]],
-          },
-          {
-            class_name: "Conv2D",
-            config: {
-              name: "block_6_project",
-              trainable: true,
-              dtype: "float32",
-              filters: 24,
-              kernel_size: [1, 1],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              kernel_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              bias_initializer: { class_name: "Zeros", config: {} },
-              kernel_regularizer: null,
-              bias_regularizer: null,
-              activity_regularizer: null,
-              kernel_constraint: null,
-              bias_constraint: null,
-            },
-            name: "block_6_project",
-            inbound_nodes: [[["block_6_depthwise_relu", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_6_project_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_6_project_BN",
-            inbound_nodes: [[["block_6_project", 0, 0, {}]]],
-          },
-          {
-            class_name: "Conv2D",
-            config: {
-              name: "block_7_expand",
-              trainable: true,
-              dtype: "float32",
-              filters: 144,
-              kernel_size: [1, 1],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              kernel_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              bias_initializer: { class_name: "Zeros", config: {} },
-              kernel_regularizer: null,
-              bias_regularizer: null,
-              activity_regularizer: null,
-              kernel_constraint: null,
-              bias_constraint: null,
-            },
-            name: "block_7_expand",
-            inbound_nodes: [[["block_6_project_BN", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_7_expand_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_7_expand_BN",
-            inbound_nodes: [[["block_7_expand", 0, 0, {}]]],
-          },
-          {
-            class_name: "ReLU",
-            config: {
-              name: "block_7_expand_relu",
-              trainable: true,
-              dtype: "float32",
-              max_value: 6.0,
-              negative_slope: 0.0,
-              threshold: 0.0,
-            },
-            name: "block_7_expand_relu",
-            inbound_nodes: [[["block_7_expand_BN", 0, 0, {}]]],
-          },
-          {
-            class_name: "DepthwiseConv2D",
-            config: {
-              name: "block_7_depthwise",
-              trainable: true,
-              dtype: "float32",
-              kernel_size: [3, 3],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              bias_initializer: { class_name: "Zeros", config: {} },
-              bias_regularizer: null,
-              activity_regularizer: null,
-              bias_constraint: null,
-              depth_multiplier: 1,
-              depthwise_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              depthwise_regularizer: null,
-              depthwise_constraint: null,
-            },
-            name: "block_7_depthwise",
-            inbound_nodes: [[["block_7_expand_relu", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_7_depthwise_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_7_depthwise_BN",
-            inbound_nodes: [[["block_7_depthwise", 0, 0, {}]]],
-          },
-          {
-            class_name: "ReLU",
-            config: {
-              name: "block_7_depthwise_relu",
-              trainable: true,
-              dtype: "float32",
-              max_value: 6.0,
-              negative_slope: 0.0,
-              threshold: 0.0,
-            },
-            name: "block_7_depthwise_relu",
-            inbound_nodes: [[["block_7_depthwise_BN", 0, 0, {}]]],
-          },
-          {
-            class_name: "Conv2D",
-            config: {
-              name: "block_7_project",
-              trainable: true,
-              dtype: "float32",
-              filters: 24,
-              kernel_size: [1, 1],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              kernel_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              bias_initializer: { class_name: "Zeros", config: {} },
-              kernel_regularizer: null,
-              bias_regularizer: null,
-              activity_regularizer: null,
-              kernel_constraint: null,
-              bias_constraint: null,
-            },
-            name: "block_7_project",
-            inbound_nodes: [[["block_7_depthwise_relu", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_7_project_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_7_project_BN",
-            inbound_nodes: [[["block_7_project", 0, 0, {}]]],
-          },
-          {
-            class_name: "Add",
-            config: { name: "block_7_add", trainable: true, dtype: "float32" },
-            name: "block_7_add",
-            inbound_nodes: [
-              [
-                ["block_6_project_BN", 0, 0, {}],
-                ["block_7_project_BN", 0, 0, {}],
-              ],
-            ],
-          },
-          {
-            class_name: "Conv2D",
-            config: {
-              name: "block_8_expand",
-              trainable: true,
-              dtype: "float32",
-              filters: 144,
-              kernel_size: [1, 1],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              kernel_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              bias_initializer: { class_name: "Zeros", config: {} },
-              kernel_regularizer: null,
-              bias_regularizer: null,
-              activity_regularizer: null,
-              kernel_constraint: null,
-              bias_constraint: null,
-            },
-            name: "block_8_expand",
-            inbound_nodes: [[["block_7_add", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_8_expand_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_8_expand_BN",
-            inbound_nodes: [[["block_8_expand", 0, 0, {}]]],
-          },
-          {
-            class_name: "ReLU",
-            config: {
-              name: "block_8_expand_relu",
-              trainable: true,
-              dtype: "float32",
-              max_value: 6.0,
-              negative_slope: 0.0,
-              threshold: 0.0,
-            },
-            name: "block_8_expand_relu",
-            inbound_nodes: [[["block_8_expand_BN", 0, 0, {}]]],
-          },
-          {
-            class_name: "DepthwiseConv2D",
-            config: {
-              name: "block_8_depthwise",
-              trainable: true,
-              dtype: "float32",
-              kernel_size: [3, 3],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              bias_initializer: { class_name: "Zeros", config: {} },
-              bias_regularizer: null,
-              activity_regularizer: null,
-              bias_constraint: null,
-              depth_multiplier: 1,
-              depthwise_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              depthwise_regularizer: null,
-              depthwise_constraint: null,
-            },
-            name: "block_8_depthwise",
-            inbound_nodes: [[["block_8_expand_relu", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_8_depthwise_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_8_depthwise_BN",
-            inbound_nodes: [[["block_8_depthwise", 0, 0, {}]]],
-          },
-          {
-            class_name: "ReLU",
-            config: {
-              name: "block_8_depthwise_relu",
-              trainable: true,
-              dtype: "float32",
-              max_value: 6.0,
-              negative_slope: 0.0,
-              threshold: 0.0,
-            },
-            name: "block_8_depthwise_relu",
-            inbound_nodes: [[["block_8_depthwise_BN", 0, 0, {}]]],
-          },
-          {
-            class_name: "Conv2D",
-            config: {
-              name: "block_8_project",
-              trainable: true,
-              dtype: "float32",
-              filters: 24,
-              kernel_size: [1, 1],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              kernel_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              bias_initializer: { class_name: "Zeros", config: {} },
-              kernel_regularizer: null,
-              bias_regularizer: null,
-              activity_regularizer: null,
-              kernel_constraint: null,
-              bias_constraint: null,
-            },
-            name: "block_8_project",
-            inbound_nodes: [[["block_8_depthwise_relu", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_8_project_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_8_project_BN",
-            inbound_nodes: [[["block_8_project", 0, 0, {}]]],
-          },
-          {
-            class_name: "Add",
-            config: { name: "block_8_add", trainable: true, dtype: "float32" },
-            name: "block_8_add",
-            inbound_nodes: [
-              [
-                ["block_7_add", 0, 0, {}],
-                ["block_8_project_BN", 0, 0, {}],
-              ],
-            ],
-          },
-          {
-            class_name: "Conv2D",
-            config: {
-              name: "block_9_expand",
-              trainable: true,
-              dtype: "float32",
-              filters: 144,
-              kernel_size: [1, 1],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              kernel_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              bias_initializer: { class_name: "Zeros", config: {} },
-              kernel_regularizer: null,
-              bias_regularizer: null,
-              activity_regularizer: null,
-              kernel_constraint: null,
-              bias_constraint: null,
-            },
-            name: "block_9_expand",
-            inbound_nodes: [[["block_8_add", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_9_expand_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_9_expand_BN",
-            inbound_nodes: [[["block_9_expand", 0, 0, {}]]],
-          },
-          {
-            class_name: "ReLU",
-            config: {
-              name: "block_9_expand_relu",
-              trainable: true,
-              dtype: "float32",
-              max_value: 6.0,
-              negative_slope: 0.0,
-              threshold: 0.0,
-            },
-            name: "block_9_expand_relu",
-            inbound_nodes: [[["block_9_expand_BN", 0, 0, {}]]],
-          },
-          {
-            class_name: "DepthwiseConv2D",
-            config: {
-              name: "block_9_depthwise",
-              trainable: true,
-              dtype: "float32",
-              kernel_size: [3, 3],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              bias_initializer: { class_name: "Zeros", config: {} },
-              bias_regularizer: null,
-              activity_regularizer: null,
-              bias_constraint: null,
-              depth_multiplier: 1,
-              depthwise_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              depthwise_regularizer: null,
-              depthwise_constraint: null,
-            },
-            name: "block_9_depthwise",
-            inbound_nodes: [[["block_9_expand_relu", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_9_depthwise_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_9_depthwise_BN",
-            inbound_nodes: [[["block_9_depthwise", 0, 0, {}]]],
-          },
-          {
-            class_name: "ReLU",
-            config: {
-              name: "block_9_depthwise_relu",
-              trainable: true,
-              dtype: "float32",
-              max_value: 6.0,
-              negative_slope: 0.0,
-              threshold: 0.0,
-            },
-            name: "block_9_depthwise_relu",
-            inbound_nodes: [[["block_9_depthwise_BN", 0, 0, {}]]],
-          },
-          {
-            class_name: "Conv2D",
-            config: {
-              name: "block_9_project",
-              trainable: true,
-              dtype: "float32",
-              filters: 24,
-              kernel_size: [1, 1],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              kernel_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              bias_initializer: { class_name: "Zeros", config: {} },
-              kernel_regularizer: null,
-              bias_regularizer: null,
-              activity_regularizer: null,
-              kernel_constraint: null,
-              bias_constraint: null,
-            },
-            name: "block_9_project",
-            inbound_nodes: [[["block_9_depthwise_relu", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_9_project_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_9_project_BN",
-            inbound_nodes: [[["block_9_project", 0, 0, {}]]],
-          },
-          {
-            class_name: "Add",
-            config: { name: "block_9_add", trainable: true, dtype: "float32" },
-            name: "block_9_add",
-            inbound_nodes: [
-              [
-                ["block_8_add", 0, 0, {}],
-                ["block_9_project_BN", 0, 0, {}],
-              ],
-            ],
-          },
-          {
-            class_name: "Conv2D",
-            config: {
-              name: "block_10_expand",
-              trainable: true,
-              dtype: "float32",
-              filters: 144,
-              kernel_size: [1, 1],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              kernel_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              bias_initializer: { class_name: "Zeros", config: {} },
-              kernel_regularizer: null,
-              bias_regularizer: null,
-              activity_regularizer: null,
-              kernel_constraint: null,
-              bias_constraint: null,
-            },
-            name: "block_10_expand",
-            inbound_nodes: [[["block_9_add", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_10_expand_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_10_expand_BN",
-            inbound_nodes: [[["block_10_expand", 0, 0, {}]]],
-          },
-          {
-            class_name: "ReLU",
-            config: {
-              name: "block_10_expand_relu",
-              trainable: true,
-              dtype: "float32",
-              max_value: 6.0,
-              negative_slope: 0.0,
-              threshold: 0.0,
-            },
-            name: "block_10_expand_relu",
-            inbound_nodes: [[["block_10_expand_BN", 0, 0, {}]]],
-          },
-          {
-            class_name: "DepthwiseConv2D",
-            config: {
-              name: "block_10_depthwise",
-              trainable: true,
-              dtype: "float32",
-              kernel_size: [3, 3],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              bias_initializer: { class_name: "Zeros", config: {} },
-              bias_regularizer: null,
-              activity_regularizer: null,
-              bias_constraint: null,
-              depth_multiplier: 1,
-              depthwise_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              depthwise_regularizer: null,
-              depthwise_constraint: null,
-            },
-            name: "block_10_depthwise",
-            inbound_nodes: [[["block_10_expand_relu", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_10_depthwise_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_10_depthwise_BN",
-            inbound_nodes: [[["block_10_depthwise", 0, 0, {}]]],
-          },
-          {
-            class_name: "ReLU",
-            config: {
-              name: "block_10_depthwise_relu",
-              trainable: true,
-              dtype: "float32",
-              max_value: 6.0,
-              negative_slope: 0.0,
-              threshold: 0.0,
-            },
-            name: "block_10_depthwise_relu",
-            inbound_nodes: [[["block_10_depthwise_BN", 0, 0, {}]]],
-          },
-          {
-            class_name: "Conv2D",
-            config: {
-              name: "block_10_project",
-              trainable: true,
-              dtype: "float32",
-              filters: 32,
-              kernel_size: [1, 1],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              kernel_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              bias_initializer: { class_name: "Zeros", config: {} },
-              kernel_regularizer: null,
-              bias_regularizer: null,
-              activity_regularizer: null,
-              kernel_constraint: null,
-              bias_constraint: null,
-            },
-            name: "block_10_project",
-            inbound_nodes: [[["block_10_depthwise_relu", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_10_project_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_10_project_BN",
-            inbound_nodes: [[["block_10_project", 0, 0, {}]]],
-          },
-          {
-            class_name: "Conv2D",
-            config: {
-              name: "block_11_expand",
-              trainable: true,
-              dtype: "float32",
-              filters: 192,
-              kernel_size: [1, 1],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              kernel_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              bias_initializer: { class_name: "Zeros", config: {} },
-              kernel_regularizer: null,
-              bias_regularizer: null,
-              activity_regularizer: null,
-              kernel_constraint: null,
-              bias_constraint: null,
-            },
-            name: "block_11_expand",
-            inbound_nodes: [[["block_10_project_BN", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_11_expand_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_11_expand_BN",
-            inbound_nodes: [[["block_11_expand", 0, 0, {}]]],
-          },
-          {
-            class_name: "ReLU",
-            config: {
-              name: "block_11_expand_relu",
-              trainable: true,
-              dtype: "float32",
-              max_value: 6.0,
-              negative_slope: 0.0,
-              threshold: 0.0,
-            },
-            name: "block_11_expand_relu",
-            inbound_nodes: [[["block_11_expand_BN", 0, 0, {}]]],
-          },
-          {
-            class_name: "DepthwiseConv2D",
-            config: {
-              name: "block_11_depthwise",
-              trainable: true,
-              dtype: "float32",
-              kernel_size: [3, 3],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              bias_initializer: { class_name: "Zeros", config: {} },
-              bias_regularizer: null,
-              activity_regularizer: null,
-              bias_constraint: null,
-              depth_multiplier: 1,
-              depthwise_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              depthwise_regularizer: null,
-              depthwise_constraint: null,
-            },
-            name: "block_11_depthwise",
-            inbound_nodes: [[["block_11_expand_relu", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_11_depthwise_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_11_depthwise_BN",
-            inbound_nodes: [[["block_11_depthwise", 0, 0, {}]]],
-          },
-          {
-            class_name: "ReLU",
-            config: {
-              name: "block_11_depthwise_relu",
-              trainable: true,
-              dtype: "float32",
-              max_value: 6.0,
-              negative_slope: 0.0,
-              threshold: 0.0,
-            },
-            name: "block_11_depthwise_relu",
-            inbound_nodes: [[["block_11_depthwise_BN", 0, 0, {}]]],
-          },
-          {
-            class_name: "Conv2D",
-            config: {
-              name: "block_11_project",
-              trainable: true,
-              dtype: "float32",
-              filters: 32,
-              kernel_size: [1, 1],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              kernel_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              bias_initializer: { class_name: "Zeros", config: {} },
-              kernel_regularizer: null,
-              bias_regularizer: null,
-              activity_regularizer: null,
-              kernel_constraint: null,
-              bias_constraint: null,
-            },
-            name: "block_11_project",
-            inbound_nodes: [[["block_11_depthwise_relu", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_11_project_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_11_project_BN",
-            inbound_nodes: [[["block_11_project", 0, 0, {}]]],
-          },
-          {
-            class_name: "Add",
-            config: { name: "block_11_add", trainable: true, dtype: "float32" },
-            name: "block_11_add",
-            inbound_nodes: [
-              [
-                ["block_10_project_BN", 0, 0, {}],
-                ["block_11_project_BN", 0, 0, {}],
-              ],
-            ],
-          },
-          {
-            class_name: "Conv2D",
-            config: {
-              name: "block_12_expand",
-              trainable: true,
-              dtype: "float32",
-              filters: 192,
-              kernel_size: [1, 1],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              kernel_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              bias_initializer: { class_name: "Zeros", config: {} },
-              kernel_regularizer: null,
-              bias_regularizer: null,
-              activity_regularizer: null,
-              kernel_constraint: null,
-              bias_constraint: null,
-            },
-            name: "block_12_expand",
-            inbound_nodes: [[["block_11_add", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_12_expand_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_12_expand_BN",
-            inbound_nodes: [[["block_12_expand", 0, 0, {}]]],
-          },
-          {
-            class_name: "ReLU",
-            config: {
-              name: "block_12_expand_relu",
-              trainable: true,
-              dtype: "float32",
-              max_value: 6.0,
-              negative_slope: 0.0,
-              threshold: 0.0,
-            },
-            name: "block_12_expand_relu",
-            inbound_nodes: [[["block_12_expand_BN", 0, 0, {}]]],
-          },
-          {
-            class_name: "DepthwiseConv2D",
-            config: {
-              name: "block_12_depthwise",
-              trainable: true,
-              dtype: "float32",
-              kernel_size: [3, 3],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              bias_initializer: { class_name: "Zeros", config: {} },
-              bias_regularizer: null,
-              activity_regularizer: null,
-              bias_constraint: null,
-              depth_multiplier: 1,
-              depthwise_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              depthwise_regularizer: null,
-              depthwise_constraint: null,
-            },
-            name: "block_12_depthwise",
-            inbound_nodes: [[["block_12_expand_relu", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_12_depthwise_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_12_depthwise_BN",
-            inbound_nodes: [[["block_12_depthwise", 0, 0, {}]]],
-          },
-          {
-            class_name: "ReLU",
-            config: {
-              name: "block_12_depthwise_relu",
-              trainable: true,
-              dtype: "float32",
-              max_value: 6.0,
-              negative_slope: 0.0,
-              threshold: 0.0,
-            },
-            name: "block_12_depthwise_relu",
-            inbound_nodes: [[["block_12_depthwise_BN", 0, 0, {}]]],
-          },
-          {
-            class_name: "Conv2D",
-            config: {
-              name: "block_12_project",
-              trainable: true,
-              dtype: "float32",
-              filters: 32,
-              kernel_size: [1, 1],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              kernel_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              bias_initializer: { class_name: "Zeros", config: {} },
-              kernel_regularizer: null,
-              bias_regularizer: null,
-              activity_regularizer: null,
-              kernel_constraint: null,
-              bias_constraint: null,
-            },
-            name: "block_12_project",
-            inbound_nodes: [[["block_12_depthwise_relu", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_12_project_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_12_project_BN",
-            inbound_nodes: [[["block_12_project", 0, 0, {}]]],
-          },
-          {
-            class_name: "Add",
-            config: { name: "block_12_add", trainable: true, dtype: "float32" },
-            name: "block_12_add",
-            inbound_nodes: [
-              [
-                ["block_11_add", 0, 0, {}],
-                ["block_12_project_BN", 0, 0, {}],
-              ],
-            ],
-          },
-          {
-            class_name: "Conv2D",
-            config: {
-              name: "block_13_expand",
-              trainable: true,
-              dtype: "float32",
-              filters: 192,
-              kernel_size: [1, 1],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              kernel_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              bias_initializer: { class_name: "Zeros", config: {} },
-              kernel_regularizer: null,
-              bias_regularizer: null,
-              activity_regularizer: null,
-              kernel_constraint: null,
-              bias_constraint: null,
-            },
-            name: "block_13_expand",
-            inbound_nodes: [[["block_12_add", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_13_expand_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_13_expand_BN",
-            inbound_nodes: [[["block_13_expand", 0, 0, {}]]],
-          },
-          {
-            class_name: "ReLU",
-            config: {
-              name: "block_13_expand_relu",
-              trainable: true,
-              dtype: "float32",
-              max_value: 6.0,
-              negative_slope: 0.0,
-              threshold: 0.0,
-            },
-            name: "block_13_expand_relu",
-            inbound_nodes: [[["block_13_expand_BN", 0, 0, {}]]],
-          },
-          {
-            class_name: "ZeroPadding2D",
-            config: {
-              name: "block_13_pad",
-              trainable: true,
-              dtype: "float32",
-              padding: [
-                [0, 1],
-                [0, 1],
-              ],
-              data_format: "channels_last",
-            },
-            name: "block_13_pad",
-            inbound_nodes: [[["block_13_expand_relu", 0, 0, {}]]],
-          },
-          {
-            class_name: "DepthwiseConv2D",
-            config: {
-              name: "block_13_depthwise",
-              trainable: true,
-              dtype: "float32",
-              kernel_size: [3, 3],
-              strides: [2, 2],
-              padding: "valid",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              bias_initializer: { class_name: "Zeros", config: {} },
-              bias_regularizer: null,
-              activity_regularizer: null,
-              bias_constraint: null,
-              depth_multiplier: 1,
-              depthwise_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              depthwise_regularizer: null,
-              depthwise_constraint: null,
-            },
-            name: "block_13_depthwise",
-            inbound_nodes: [[["block_13_pad", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_13_depthwise_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_13_depthwise_BN",
-            inbound_nodes: [[["block_13_depthwise", 0, 0, {}]]],
-          },
-          {
-            class_name: "ReLU",
-            config: {
-              name: "block_13_depthwise_relu",
-              trainable: true,
-              dtype: "float32",
-              max_value: 6.0,
-              negative_slope: 0.0,
-              threshold: 0.0,
-            },
-            name: "block_13_depthwise_relu",
-            inbound_nodes: [[["block_13_depthwise_BN", 0, 0, {}]]],
-          },
-          {
-            class_name: "Conv2D",
-            config: {
-              name: "block_13_project",
-              trainable: true,
-              dtype: "float32",
-              filters: 56,
-              kernel_size: [1, 1],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              kernel_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              bias_initializer: { class_name: "Zeros", config: {} },
-              kernel_regularizer: null,
-              bias_regularizer: null,
-              activity_regularizer: null,
-              kernel_constraint: null,
-              bias_constraint: null,
-            },
-            name: "block_13_project",
-            inbound_nodes: [[["block_13_depthwise_relu", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_13_project_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_13_project_BN",
-            inbound_nodes: [[["block_13_project", 0, 0, {}]]],
-          },
-          {
-            class_name: "Conv2D",
-            config: {
-              name: "block_14_expand",
-              trainable: true,
-              dtype: "float32",
-              filters: 336,
-              kernel_size: [1, 1],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              kernel_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              bias_initializer: { class_name: "Zeros", config: {} },
-              kernel_regularizer: null,
-              bias_regularizer: null,
-              activity_regularizer: null,
-              kernel_constraint: null,
-              bias_constraint: null,
-            },
-            name: "block_14_expand",
-            inbound_nodes: [[["block_13_project_BN", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_14_expand_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_14_expand_BN",
-            inbound_nodes: [[["block_14_expand", 0, 0, {}]]],
-          },
-          {
-            class_name: "ReLU",
-            config: {
-              name: "block_14_expand_relu",
-              trainable: true,
-              dtype: "float32",
-              max_value: 6.0,
-              negative_slope: 0.0,
-              threshold: 0.0,
-            },
-            name: "block_14_expand_relu",
-            inbound_nodes: [[["block_14_expand_BN", 0, 0, {}]]],
-          },
-          {
-            class_name: "DepthwiseConv2D",
-            config: {
-              name: "block_14_depthwise",
-              trainable: true,
-              dtype: "float32",
-              kernel_size: [3, 3],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              bias_initializer: { class_name: "Zeros", config: {} },
-              bias_regularizer: null,
-              activity_regularizer: null,
-              bias_constraint: null,
-              depth_multiplier: 1,
-              depthwise_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              depthwise_regularizer: null,
-              depthwise_constraint: null,
-            },
-            name: "block_14_depthwise",
-            inbound_nodes: [[["block_14_expand_relu", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_14_depthwise_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_14_depthwise_BN",
-            inbound_nodes: [[["block_14_depthwise", 0, 0, {}]]],
-          },
-          {
-            class_name: "ReLU",
-            config: {
-              name: "block_14_depthwise_relu",
-              trainable: true,
-              dtype: "float32",
-              max_value: 6.0,
-              negative_slope: 0.0,
-              threshold: 0.0,
-            },
-            name: "block_14_depthwise_relu",
-            inbound_nodes: [[["block_14_depthwise_BN", 0, 0, {}]]],
-          },
-          {
-            class_name: "Conv2D",
-            config: {
-              name: "block_14_project",
-              trainable: true,
-              dtype: "float32",
-              filters: 56,
-              kernel_size: [1, 1],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              kernel_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              bias_initializer: { class_name: "Zeros", config: {} },
-              kernel_regularizer: null,
-              bias_regularizer: null,
-              activity_regularizer: null,
-              kernel_constraint: null,
-              bias_constraint: null,
-            },
-            name: "block_14_project",
-            inbound_nodes: [[["block_14_depthwise_relu", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_14_project_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_14_project_BN",
-            inbound_nodes: [[["block_14_project", 0, 0, {}]]],
-          },
-          {
-            class_name: "Add",
-            config: { name: "block_14_add", trainable: true, dtype: "float32" },
-            name: "block_14_add",
-            inbound_nodes: [
-              [
-                ["block_13_project_BN", 0, 0, {}],
-                ["block_14_project_BN", 0, 0, {}],
-              ],
-            ],
-          },
-          {
-            class_name: "Conv2D",
-            config: {
-              name: "block_15_expand",
-              trainable: true,
-              dtype: "float32",
-              filters: 336,
-              kernel_size: [1, 1],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              kernel_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              bias_initializer: { class_name: "Zeros", config: {} },
-              kernel_regularizer: null,
-              bias_regularizer: null,
-              activity_regularizer: null,
-              kernel_constraint: null,
-              bias_constraint: null,
-            },
-            name: "block_15_expand",
-            inbound_nodes: [[["block_14_add", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_15_expand_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_15_expand_BN",
-            inbound_nodes: [[["block_15_expand", 0, 0, {}]]],
-          },
-          {
-            class_name: "ReLU",
-            config: {
-              name: "block_15_expand_relu",
-              trainable: true,
-              dtype: "float32",
-              max_value: 6.0,
-              negative_slope: 0.0,
-              threshold: 0.0,
-            },
-            name: "block_15_expand_relu",
-            inbound_nodes: [[["block_15_expand_BN", 0, 0, {}]]],
-          },
-          {
-            class_name: "DepthwiseConv2D",
-            config: {
-              name: "block_15_depthwise",
-              trainable: true,
-              dtype: "float32",
-              kernel_size: [3, 3],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              bias_initializer: { class_name: "Zeros", config: {} },
-              bias_regularizer: null,
-              activity_regularizer: null,
-              bias_constraint: null,
-              depth_multiplier: 1,
-              depthwise_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              depthwise_regularizer: null,
-              depthwise_constraint: null,
-            },
-            name: "block_15_depthwise",
-            inbound_nodes: [[["block_15_expand_relu", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_15_depthwise_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_15_depthwise_BN",
-            inbound_nodes: [[["block_15_depthwise", 0, 0, {}]]],
-          },
-          {
-            class_name: "ReLU",
-            config: {
-              name: "block_15_depthwise_relu",
-              trainable: true,
-              dtype: "float32",
-              max_value: 6.0,
-              negative_slope: 0.0,
-              threshold: 0.0,
-            },
-            name: "block_15_depthwise_relu",
-            inbound_nodes: [[["block_15_depthwise_BN", 0, 0, {}]]],
-          },
-          {
-            class_name: "Conv2D",
-            config: {
-              name: "block_15_project",
-              trainable: true,
-              dtype: "float32",
-              filters: 56,
-              kernel_size: [1, 1],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              kernel_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              bias_initializer: { class_name: "Zeros", config: {} },
-              kernel_regularizer: null,
-              bias_regularizer: null,
-              activity_regularizer: null,
-              kernel_constraint: null,
-              bias_constraint: null,
-            },
-            name: "block_15_project",
-            inbound_nodes: [[["block_15_depthwise_relu", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_15_project_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_15_project_BN",
-            inbound_nodes: [[["block_15_project", 0, 0, {}]]],
-          },
-          {
-            class_name: "Add",
-            config: { name: "block_15_add", trainable: true, dtype: "float32" },
-            name: "block_15_add",
-            inbound_nodes: [
-              [
-                ["block_14_add", 0, 0, {}],
-                ["block_15_project_BN", 0, 0, {}],
-              ],
-            ],
-          },
-          {
-            class_name: "Conv2D",
-            config: {
-              name: "block_16_expand",
-              trainable: true,
-              dtype: "float32",
-              filters: 336,
-              kernel_size: [1, 1],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              kernel_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              bias_initializer: { class_name: "Zeros", config: {} },
-              kernel_regularizer: null,
-              bias_regularizer: null,
-              activity_regularizer: null,
-              kernel_constraint: null,
-              bias_constraint: null,
-            },
-            name: "block_16_expand",
-            inbound_nodes: [[["block_15_add", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_16_expand_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_16_expand_BN",
-            inbound_nodes: [[["block_16_expand", 0, 0, {}]]],
-          },
-          {
-            class_name: "ReLU",
-            config: {
-              name: "block_16_expand_relu",
-              trainable: true,
-              dtype: "float32",
-              max_value: 6.0,
-              negative_slope: 0.0,
-              threshold: 0.0,
-            },
-            name: "block_16_expand_relu",
-            inbound_nodes: [[["block_16_expand_BN", 0, 0, {}]]],
-          },
-          {
-            class_name: "DepthwiseConv2D",
-            config: {
-              name: "block_16_depthwise",
-              trainable: true,
-              dtype: "float32",
-              kernel_size: [3, 3],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              bias_initializer: { class_name: "Zeros", config: {} },
-              bias_regularizer: null,
-              activity_regularizer: null,
-              bias_constraint: null,
-              depth_multiplier: 1,
-              depthwise_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              depthwise_regularizer: null,
-              depthwise_constraint: null,
-            },
-            name: "block_16_depthwise",
-            inbound_nodes: [[["block_16_expand_relu", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_16_depthwise_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_16_depthwise_BN",
-            inbound_nodes: [[["block_16_depthwise", 0, 0, {}]]],
-          },
-          {
-            class_name: "ReLU",
-            config: {
-              name: "block_16_depthwise_relu",
-              trainable: true,
-              dtype: "float32",
-              max_value: 6.0,
-              negative_slope: 0.0,
-              threshold: 0.0,
-            },
-            name: "block_16_depthwise_relu",
-            inbound_nodes: [[["block_16_depthwise_BN", 0, 0, {}]]],
-          },
-          {
-            class_name: "Conv2D",
-            config: {
-              name: "block_16_project",
-              trainable: true,
-              dtype: "float32",
-              filters: 112,
-              kernel_size: [1, 1],
-              strides: [1, 1],
-              padding: "same",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              kernel_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              bias_initializer: { class_name: "Zeros", config: {} },
-              kernel_regularizer: null,
-              bias_regularizer: null,
-              activity_regularizer: null,
-              kernel_constraint: null,
-              bias_constraint: null,
-            },
-            name: "block_16_project",
-            inbound_nodes: [[["block_16_depthwise_relu", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "block_16_project_BN",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "block_16_project_BN",
-            inbound_nodes: [[["block_16_project", 0, 0, {}]]],
-          },
-          {
-            class_name: "Conv2D",
-            config: {
-              name: "Conv_1",
-              trainable: true,
-              dtype: "float32",
-              filters: 1280,
-              kernel_size: [1, 1],
-              strides: [1, 1],
-              padding: "valid",
-              data_format: "channels_last",
-              dilation_rate: [1, 1],
-              groups: 1,
-              activation: "linear",
-              use_bias: false,
-              kernel_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              bias_initializer: { class_name: "Zeros", config: {} },
-              kernel_regularizer: null,
-              bias_regularizer: null,
-              activity_regularizer: null,
-              kernel_constraint: null,
-              bias_constraint: null,
-            },
-            name: "Conv_1",
-            inbound_nodes: [[["block_16_project_BN", 0, 0, {}]]],
-          },
-          {
-            class_name: "BatchNormalization",
-            config: {
-              name: "Conv_1_bn",
-              trainable: true,
-              dtype: "float32",
-              axis: [3],
-              momentum: 0.999,
-              epsilon: 0.001,
-              center: true,
-              scale: true,
-              beta_initializer: { class_name: "Zeros", config: {} },
-              gamma_initializer: { class_name: "Ones", config: {} },
-              moving_mean_initializer: { class_name: "Zeros", config: {} },
-              moving_variance_initializer: { class_name: "Ones", config: {} },
-              beta_regularizer: null,
-              gamma_regularizer: null,
-              beta_constraint: null,
-              gamma_constraint: null,
-            },
-            name: "Conv_1_bn",
-            inbound_nodes: [[["Conv_1", 0, 0, {}]]],
-          },
-          {
-            class_name: "ReLU",
-            config: {
-              name: "out_relu",
-              trainable: true,
-              dtype: "float32",
-              max_value: 6.0,
-              negative_slope: 0.0,
-              threshold: 0.0,
-            },
-            name: "out_relu",
-            inbound_nodes: [[["Conv_1_bn", 0, 0, {}]]],
-          },
-          {
-            class_name: "GlobalAveragePooling2D",
-            config: {
-              name: "global_average_pooling2d",
-              trainable: true,
-              dtype: "float32",
-              data_format: "channels_last",
-              keepdims: false,
-            },
-            name: "global_average_pooling2d",
-            inbound_nodes: [[["out_relu", 0, 0, {}]]],
-          },
-          {
-            class_name: "Dense",
-            config: {
-              name: "dense",
-              trainable: true,
-              dtype: "float32",
-              units: 1024,
-              activation: "relu",
-              use_bias: true,
-              kernel_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              bias_initializer: { class_name: "Zeros", config: {} },
-              kernel_regularizer: null,
-              bias_regularizer: null,
-              activity_regularizer: null,
-              kernel_constraint: null,
-              bias_constraint: null,
-            },
-            name: "dense",
-            inbound_nodes: [[["global_average_pooling2d", 0, 0, {}]]],
-          },
-          {
-            class_name: "Dense",
-            config: {
-              name: "dense_1",
-              trainable: true,
-              dtype: "float32",
-              units: 2,
-              activation: "softmax",
-              use_bias: true,
-              kernel_initializer: {
-                class_name: "GlorotUniform",
-                config: { seed: null },
-              },
-              bias_initializer: { class_name: "Zeros", config: {} },
-              kernel_regularizer: null,
-              bias_regularizer: null,
-              activity_regularizer: null,
-              kernel_constraint: null,
-              bias_constraint: null,
-            },
-            name: "dense_1",
-            inbound_nodes: [[["dense", 0, 0, {}]]],
-          },
-        ],
-        input_layers: [["input_1", 0, 0]],
-        output_layers: [["dense_1", 0, 0]],
-      },
-    },
-  },
-  weightsManifest: [
-    {
-      paths: ["group1-shard1of2.bin", "group1-shard2of2.bin"],
-      weights: [
-        { name: "Conv1/kernel", shape: [3, 3, 3, 16], dtype: "float32" },
-        { name: "Conv_1/kernel", shape: [1, 1, 112, 1280], dtype: "float32" },
-        { name: "Conv_1_bn/gamma", shape: [1280], dtype: "float32" },
-        { name: "Conv_1_bn/beta", shape: [1280], dtype: "float32" },
-        { name: "Conv_1_bn/moving_mean", shape: [1280], dtype: "float32" },
-        { name: "Conv_1_bn/moving_variance", shape: [1280], dtype: "float32" },
-        {
-          name: "block_10_depthwise/depthwise_kernel",
-          shape: [3, 3, 144, 1],
-          dtype: "float32",
-        },
-        { name: "block_10_depthwise_BN/gamma", shape: [144], dtype: "float32" },
-        { name: "block_10_depthwise_BN/beta", shape: [144], dtype: "float32" },
-        {
-          name: "block_10_depthwise_BN/moving_mean",
-          shape: [144],
-          dtype: "float32",
-        },
-        {
-          name: "block_10_depthwise_BN/moving_variance",
-          shape: [144],
-          dtype: "float32",
-        },
-        {
-          name: "block_10_expand/kernel",
-          shape: [1, 1, 24, 144],
-          dtype: "float32",
-        },
-        { name: "block_10_expand_BN/gamma", shape: [144], dtype: "float32" },
-        { name: "block_10_expand_BN/beta", shape: [144], dtype: "float32" },
-        {
-          name: "block_10_expand_BN/moving_mean",
-          shape: [144],
-          dtype: "float32",
-        },
-        {
-          name: "block_10_expand_BN/moving_variance",
-          shape: [144],
-          dtype: "float32",
-        },
-        {
-          name: "block_10_project/kernel",
-          shape: [1, 1, 144, 32],
-          dtype: "float32",
-        },
-        { name: "block_10_project_BN/gamma", shape: [32], dtype: "float32" },
-        { name: "block_10_project_BN/beta", shape: [32], dtype: "float32" },
-        {
-          name: "block_10_project_BN/moving_mean",
-          shape: [32],
-          dtype: "float32",
-        },
-        {
-          name: "block_10_project_BN/moving_variance",
-          shape: [32],
-          dtype: "float32",
-        },
-        {
-          name: "block_11_depthwise/depthwise_kernel",
-          shape: [3, 3, 192, 1],
-          dtype: "float32",
-        },
-        { name: "block_11_depthwise_BN/gamma", shape: [192], dtype: "float32" },
-        { name: "block_11_depthwise_BN/beta", shape: [192], dtype: "float32" },
-        {
-          name: "block_11_depthwise_BN/moving_mean",
-          shape: [192],
-          dtype: "float32",
-        },
-        {
-          name: "block_11_depthwise_BN/moving_variance",
-          shape: [192],
-          dtype: "float32",
-        },
-        {
-          name: "block_11_expand/kernel",
-          shape: [1, 1, 32, 192],
-          dtype: "float32",
-        },
-        { name: "block_11_expand_BN/gamma", shape: [192], dtype: "float32" },
-        { name: "block_11_expand_BN/beta", shape: [192], dtype: "float32" },
-        {
-          name: "block_11_expand_BN/moving_mean",
-          shape: [192],
-          dtype: "float32",
-        },
-        {
-          name: "block_11_expand_BN/moving_variance",
-          shape: [192],
-          dtype: "float32",
-        },
-        {
-          name: "block_11_project/kernel",
-          shape: [1, 1, 192, 32],
-          dtype: "float32",
-        },
-        { name: "block_11_project_BN/gamma", shape: [32], dtype: "float32" },
-        { name: "block_11_project_BN/beta", shape: [32], dtype: "float32" },
-        {
-          name: "block_11_project_BN/moving_mean",
-          shape: [32],
-          dtype: "float32",
-        },
-        {
-          name: "block_11_project_BN/moving_variance",
-          shape: [32],
-          dtype: "float32",
-        },
-        {
-          name: "block_12_depthwise/depthwise_kernel",
-          shape: [3, 3, 192, 1],
-          dtype: "float32",
-        },
-        { name: "block_12_depthwise_BN/gamma", shape: [192], dtype: "float32" },
-        { name: "block_12_depthwise_BN/beta", shape: [192], dtype: "float32" },
-        {
-          name: "block_12_depthwise_BN/moving_mean",
-          shape: [192],
-          dtype: "float32",
-        },
-        {
-          name: "block_12_depthwise_BN/moving_variance",
-          shape: [192],
-          dtype: "float32",
-        },
-        {
-          name: "block_12_expand/kernel",
-          shape: [1, 1, 32, 192],
-          dtype: "float32",
-        },
-        { name: "block_12_expand_BN/gamma", shape: [192], dtype: "float32" },
-        { name: "block_12_expand_BN/beta", shape: [192], dtype: "float32" },
-        {
-          name: "block_12_expand_BN/moving_mean",
-          shape: [192],
-          dtype: "float32",
-        },
-        {
-          name: "block_12_expand_BN/moving_variance",
-          shape: [192],
-          dtype: "float32",
-        },
-        {
-          name: "block_12_project/kernel",
-          shape: [1, 1, 192, 32],
-          dtype: "float32",
-        },
-        { name: "block_12_project_BN/gamma", shape: [32], dtype: "float32" },
-        { name: "block_12_project_BN/beta", shape: [32], dtype: "float32" },
-        {
-          name: "block_12_project_BN/moving_mean",
-          shape: [32],
-          dtype: "float32",
-        },
-        {
-          name: "block_12_project_BN/moving_variance",
-          shape: [32],
-          dtype: "float32",
-        },
-        {
-          name: "block_13_depthwise/depthwise_kernel",
-          shape: [3, 3, 192, 1],
-          dtype: "float32",
-        },
-        { name: "block_13_depthwise_BN/gamma", shape: [192], dtype: "float32" },
-        { name: "block_13_depthwise_BN/beta", shape: [192], dtype: "float32" },
-        {
-          name: "block_13_depthwise_BN/moving_mean",
-          shape: [192],
-          dtype: "float32",
-        },
-        {
-          name: "block_13_depthwise_BN/moving_variance",
-          shape: [192],
-          dtype: "float32",
-        },
-        {
-          name: "block_13_expand/kernel",
-          shape: [1, 1, 32, 192],
-          dtype: "float32",
-        },
-        { name: "block_13_expand_BN/gamma", shape: [192], dtype: "float32" },
-        { name: "block_13_expand_BN/beta", shape: [192], dtype: "float32" },
-        {
-          name: "block_13_expand_BN/moving_mean",
-          shape: [192],
-          dtype: "float32",
-        },
-        {
-          name: "block_13_expand_BN/moving_variance",
-          shape: [192],
-          dtype: "float32",
-        },
-        {
-          name: "block_13_project/kernel",
-          shape: [1, 1, 192, 56],
-          dtype: "float32",
-        },
-        { name: "block_13_project_BN/gamma", shape: [56], dtype: "float32" },
-        { name: "block_13_project_BN/beta", shape: [56], dtype: "float32" },
-        {
-          name: "block_13_project_BN/moving_mean",
-          shape: [56],
-          dtype: "float32",
-        },
-        {
-          name: "block_13_project_BN/moving_variance",
-          shape: [56],
-          dtype: "float32",
-        },
-        {
-          name: "block_14_depthwise/depthwise_kernel",
-          shape: [3, 3, 336, 1],
-          dtype: "float32",
-        },
-        { name: "block_14_depthwise_BN/gamma", shape: [336], dtype: "float32" },
-        { name: "block_14_depthwise_BN/beta", shape: [336], dtype: "float32" },
-        {
-          name: "block_14_depthwise_BN/moving_mean",
-          shape: [336],
-          dtype: "float32",
-        },
-        {
-          name: "block_14_depthwise_BN/moving_variance",
-          shape: [336],
-          dtype: "float32",
-        },
-        {
-          name: "block_14_expand/kernel",
-          shape: [1, 1, 56, 336],
-          dtype: "float32",
-        },
-        { name: "block_14_expand_BN/gamma", shape: [336], dtype: "float32" },
-        { name: "block_14_expand_BN/beta", shape: [336], dtype: "float32" },
-        {
-          name: "block_14_expand_BN/moving_mean",
-          shape: [336],
-          dtype: "float32",
-        },
-        {
-          name: "block_14_expand_BN/moving_variance",
-          shape: [336],
-          dtype: "float32",
-        },
-        {
-          name: "block_14_project/kernel",
-          shape: [1, 1, 336, 56],
-          dtype: "float32",
-        },
-        { name: "block_14_project_BN/gamma", shape: [56], dtype: "float32" },
-        { name: "block_14_project_BN/beta", shape: [56], dtype: "float32" },
-        {
-          name: "block_14_project_BN/moving_mean",
-          shape: [56],
-          dtype: "float32",
-        },
-        {
-          name: "block_14_project_BN/moving_variance",
-          shape: [56],
-          dtype: "float32",
-        },
-        {
-          name: "block_15_depthwise/depthwise_kernel",
-          shape: [3, 3, 336, 1],
-          dtype: "float32",
-        },
-        { name: "block_15_depthwise_BN/gamma", shape: [336], dtype: "float32" },
-        { name: "block_15_depthwise_BN/beta", shape: [336], dtype: "float32" },
-        {
-          name: "block_15_depthwise_BN/moving_mean",
-          shape: [336],
-          dtype: "float32",
-        },
-        {
-          name: "block_15_depthwise_BN/moving_variance",
-          shape: [336],
-          dtype: "float32",
-        },
-        {
-          name: "block_15_expand/kernel",
-          shape: [1, 1, 56, 336],
-          dtype: "float32",
-        },
-        { name: "block_15_expand_BN/gamma", shape: [336], dtype: "float32" },
-        { name: "block_15_expand_BN/beta", shape: [336], dtype: "float32" },
-        {
-          name: "block_15_expand_BN/moving_mean",
-          shape: [336],
-          dtype: "float32",
-        },
-        {
-          name: "block_15_expand_BN/moving_variance",
-          shape: [336],
-          dtype: "float32",
-        },
-        {
-          name: "block_15_project/kernel",
-          shape: [1, 1, 336, 56],
-          dtype: "float32",
-        },
-        { name: "block_15_project_BN/gamma", shape: [56], dtype: "float32" },
-        { name: "block_15_project_BN/beta", shape: [56], dtype: "float32" },
-        {
-          name: "block_15_project_BN/moving_mean",
-          shape: [56],
-          dtype: "float32",
-        },
-        {
-          name: "block_15_project_BN/moving_variance",
-          shape: [56],
-          dtype: "float32",
-        },
-        {
-          name: "block_16_depthwise/depthwise_kernel",
-          shape: [3, 3, 336, 1],
-          dtype: "float32",
-        },
-        { name: "block_16_depthwise_BN/gamma", shape: [336], dtype: "float32" },
-        { name: "block_16_depthwise_BN/beta", shape: [336], dtype: "float32" },
-        {
-          name: "block_16_depthwise_BN/moving_mean",
-          shape: [336],
-          dtype: "float32",
-        },
-        {
-          name: "block_16_depthwise_BN/moving_variance",
-          shape: [336],
-          dtype: "float32",
-        },
-        {
-          name: "block_16_expand/kernel",
-          shape: [1, 1, 56, 336],
-          dtype: "float32",
-        },
-        { name: "block_16_expand_BN/gamma", shape: [336], dtype: "float32" },
-        { name: "block_16_expand_BN/beta", shape: [336], dtype: "float32" },
-        {
-          name: "block_16_expand_BN/moving_mean",
-          shape: [336],
-          dtype: "float32",
-        },
-        {
-          name: "block_16_expand_BN/moving_variance",
-          shape: [336],
-          dtype: "float32",
-        },
-        {
-          name: "block_16_project/kernel",
-          shape: [1, 1, 336, 112],
-          dtype: "float32",
-        },
-        { name: "block_16_project_BN/gamma", shape: [112], dtype: "float32" },
-        { name: "block_16_project_BN/beta", shape: [112], dtype: "float32" },
-        {
-          name: "block_16_project_BN/moving_mean",
-          shape: [112],
-          dtype: "float32",
-        },
-        {
-          name: "block_16_project_BN/moving_variance",
-          shape: [112],
-          dtype: "float32",
-        },
-        {
-          name: "block_1_depthwise/depthwise_kernel",
-          shape: [3, 3, 48, 1],
-          dtype: "float32",
-        },
-        { name: "block_1_depthwise_BN/gamma", shape: [48], dtype: "float32" },
-        { name: "block_1_depthwise_BN/beta", shape: [48], dtype: "float32" },
-        {
-          name: "block_1_depthwise_BN/moving_mean",
-          shape: [48],
-          dtype: "float32",
-        },
-        {
-          name: "block_1_depthwise_BN/moving_variance",
-          shape: [48],
-          dtype: "float32",
-        },
-        {
-          name: "block_1_expand/kernel",
-          shape: [1, 1, 8, 48],
-          dtype: "float32",
-        },
-        { name: "block_1_expand_BN/gamma", shape: [48], dtype: "float32" },
-        { name: "block_1_expand_BN/beta", shape: [48], dtype: "float32" },
-        {
-          name: "block_1_expand_BN/moving_mean",
-          shape: [48],
-          dtype: "float32",
-        },
-        {
-          name: "block_1_expand_BN/moving_variance",
-          shape: [48],
-          dtype: "float32",
-        },
-        {
-          name: "block_1_project/kernel",
-          shape: [1, 1, 48, 8],
-          dtype: "float32",
-        },
-        { name: "block_1_project_BN/gamma", shape: [8], dtype: "float32" },
-        { name: "block_1_project_BN/beta", shape: [8], dtype: "float32" },
-        {
-          name: "block_1_project_BN/moving_mean",
-          shape: [8],
-          dtype: "float32",
-        },
-        {
-          name: "block_1_project_BN/moving_variance",
-          shape: [8],
-          dtype: "float32",
-        },
-        {
-          name: "block_2_depthwise/depthwise_kernel",
-          shape: [3, 3, 48, 1],
-          dtype: "float32",
-        },
-        { name: "block_2_depthwise_BN/gamma", shape: [48], dtype: "float32" },
-        { name: "block_2_depthwise_BN/beta", shape: [48], dtype: "float32" },
-        {
-          name: "block_2_depthwise_BN/moving_mean",
-          shape: [48],
-          dtype: "float32",
-        },
-        {
-          name: "block_2_depthwise_BN/moving_variance",
-          shape: [48],
-          dtype: "float32",
-        },
-        {
-          name: "block_2_expand/kernel",
-          shape: [1, 1, 8, 48],
-          dtype: "float32",
-        },
-        { name: "block_2_expand_BN/gamma", shape: [48], dtype: "float32" },
-        { name: "block_2_expand_BN/beta", shape: [48], dtype: "float32" },
-        {
-          name: "block_2_expand_BN/moving_mean",
-          shape: [48],
-          dtype: "float32",
-        },
-        {
-          name: "block_2_expand_BN/moving_variance",
-          shape: [48],
-          dtype: "float32",
-        },
-        {
-          name: "block_2_project/kernel",
-          shape: [1, 1, 48, 8],
-          dtype: "float32",
-        },
-        { name: "block_2_project_BN/gamma", shape: [8], dtype: "float32" },
-        { name: "block_2_project_BN/beta", shape: [8], dtype: "float32" },
-        {
-          name: "block_2_project_BN/moving_mean",
-          shape: [8],
-          dtype: "float32",
-        },
-        {
-          name: "block_2_project_BN/moving_variance",
-          shape: [8],
-          dtype: "float32",
-        },
-        {
-          name: "block_3_depthwise/depthwise_kernel",
-          shape: [3, 3, 48, 1],
-          dtype: "float32",
-        },
-        { name: "block_3_depthwise_BN/gamma", shape: [48], dtype: "float32" },
-        { name: "block_3_depthwise_BN/beta", shape: [48], dtype: "float32" },
-        {
-          name: "block_3_depthwise_BN/moving_mean",
-          shape: [48],
-          dtype: "float32",
-        },
-        {
-          name: "block_3_depthwise_BN/moving_variance",
-          shape: [48],
-          dtype: "float32",
-        },
-        {
-          name: "block_3_expand/kernel",
-          shape: [1, 1, 8, 48],
-          dtype: "float32",
-        },
-        { name: "block_3_expand_BN/gamma", shape: [48], dtype: "float32" },
-        { name: "block_3_expand_BN/beta", shape: [48], dtype: "float32" },
-        {
-          name: "block_3_expand_BN/moving_mean",
-          shape: [48],
-          dtype: "float32",
-        },
-        {
-          name: "block_3_expand_BN/moving_variance",
-          shape: [48],
-          dtype: "float32",
-        },
-        {
-          name: "block_3_project/kernel",
-          shape: [1, 1, 48, 16],
-          dtype: "float32",
-        },
-        { name: "block_3_project_BN/gamma", shape: [16], dtype: "float32" },
-        { name: "block_3_project_BN/beta", shape: [16], dtype: "float32" },
-        {
-          name: "block_3_project_BN/moving_mean",
-          shape: [16],
-          dtype: "float32",
-        },
-        {
-          name: "block_3_project_BN/moving_variance",
-          shape: [16],
-          dtype: "float32",
-        },
-        {
-          name: "block_4_depthwise/depthwise_kernel",
-          shape: [3, 3, 96, 1],
-          dtype: "float32",
-        },
-        { name: "block_4_depthwise_BN/gamma", shape: [96], dtype: "float32" },
-        { name: "block_4_depthwise_BN/beta", shape: [96], dtype: "float32" },
-        {
-          name: "block_4_depthwise_BN/moving_mean",
-          shape: [96],
-          dtype: "float32",
-        },
-        {
-          name: "block_4_depthwise_BN/moving_variance",
-          shape: [96],
-          dtype: "float32",
-        },
-        {
-          name: "block_4_expand/kernel",
-          shape: [1, 1, 16, 96],
-          dtype: "float32",
-        },
-        { name: "block_4_expand_BN/gamma", shape: [96], dtype: "float32" },
-        { name: "block_4_expand_BN/beta", shape: [96], dtype: "float32" },
-        {
-          name: "block_4_expand_BN/moving_mean",
-          shape: [96],
-          dtype: "float32",
-        },
-        {
-          name: "block_4_expand_BN/moving_variance",
-          shape: [96],
-          dtype: "float32",
-        },
-        {
-          name: "block_4_project/kernel",
-          shape: [1, 1, 96, 16],
-          dtype: "float32",
-        },
-        { name: "block_4_project_BN/gamma", shape: [16], dtype: "float32" },
-        { name: "block_4_project_BN/beta", shape: [16], dtype: "float32" },
-        {
-          name: "block_4_project_BN/moving_mean",
-          shape: [16],
-          dtype: "float32",
-        },
-        {
-          name: "block_4_project_BN/moving_variance",
-          shape: [16],
-          dtype: "float32",
-        },
-        {
-          name: "block_5_depthwise/depthwise_kernel",
-          shape: [3, 3, 96, 1],
-          dtype: "float32",
-        },
-        { name: "block_5_depthwise_BN/gamma", shape: [96], dtype: "float32" },
-        { name: "block_5_depthwise_BN/beta", shape: [96], dtype: "float32" },
-        {
-          name: "block_5_depthwise_BN/moving_mean",
-          shape: [96],
-          dtype: "float32",
-        },
-        {
-          name: "block_5_depthwise_BN/moving_variance",
-          shape: [96],
-          dtype: "float32",
-        },
-        {
-          name: "block_5_expand/kernel",
-          shape: [1, 1, 16, 96],
-          dtype: "float32",
-        },
-        { name: "block_5_expand_BN/gamma", shape: [96], dtype: "float32" },
-        { name: "block_5_expand_BN/beta", shape: [96], dtype: "float32" },
-        {
-          name: "block_5_expand_BN/moving_mean",
-          shape: [96],
-          dtype: "float32",
-        },
-        {
-          name: "block_5_expand_BN/moving_variance",
-          shape: [96],
-          dtype: "float32",
-        },
-        {
-          name: "block_5_project/kernel",
-          shape: [1, 1, 96, 16],
-          dtype: "float32",
-        },
-        { name: "block_5_project_BN/gamma", shape: [16], dtype: "float32" },
-        { name: "block_5_project_BN/beta", shape: [16], dtype: "float32" },
-        {
-          name: "block_5_project_BN/moving_mean",
-          shape: [16],
-          dtype: "float32",
-        },
-        {
-          name: "block_5_project_BN/moving_variance",
-          shape: [16],
-          dtype: "float32",
-        },
-        {
-          name: "block_6_depthwise/depthwise_kernel",
-          shape: [3, 3, 96, 1],
-          dtype: "float32",
-        },
-        { name: "block_6_depthwise_BN/gamma", shape: [96], dtype: "float32" },
-        { name: "block_6_depthwise_BN/beta", shape: [96], dtype: "float32" },
-        {
-          name: "block_6_depthwise_BN/moving_mean",
-          shape: [96],
-          dtype: "float32",
-        },
-        {
-          name: "block_6_depthwise_BN/moving_variance",
-          shape: [96],
-          dtype: "float32",
-        },
-        {
-          name: "block_6_expand/kernel",
-          shape: [1, 1, 16, 96],
-          dtype: "float32",
-        },
-        { name: "block_6_expand_BN/gamma", shape: [96], dtype: "float32" },
-        { name: "block_6_expand_BN/beta", shape: [96], dtype: "float32" },
-        {
-          name: "block_6_expand_BN/moving_mean",
-          shape: [96],
-          dtype: "float32",
-        },
-        {
-          name: "block_6_expand_BN/moving_variance",
-          shape: [96],
-          dtype: "float32",
-        },
-        {
-          name: "block_6_project/kernel",
-          shape: [1, 1, 96, 24],
-          dtype: "float32",
-        },
-        { name: "block_6_project_BN/gamma", shape: [24], dtype: "float32" },
-        { name: "block_6_project_BN/beta", shape: [24], dtype: "float32" },
-        {
-          name: "block_6_project_BN/moving_mean",
-          shape: [24],
-          dtype: "float32",
-        },
-        {
-          name: "block_6_project_BN/moving_variance",
-          shape: [24],
-          dtype: "float32",
-        },
-        {
-          name: "block_7_depthwise/depthwise_kernel",
-          shape: [3, 3, 144, 1],
-          dtype: "float32",
-        },
-        { name: "block_7_depthwise_BN/gamma", shape: [144], dtype: "float32" },
-        { name: "block_7_depthwise_BN/beta", shape: [144], dtype: "float32" },
-        {
-          name: "block_7_depthwise_BN/moving_mean",
-          shape: [144],
-          dtype: "float32",
-        },
-        {
-          name: "block_7_depthwise_BN/moving_variance",
-          shape: [144],
-          dtype: "float32",
-        },
-        {
-          name: "block_7_expand/kernel",
-          shape: [1, 1, 24, 144],
-          dtype: "float32",
-        },
-        { name: "block_7_expand_BN/gamma", shape: [144], dtype: "float32" },
-        { name: "block_7_expand_BN/beta", shape: [144], dtype: "float32" },
-        {
-          name: "block_7_expand_BN/moving_mean",
-          shape: [144],
-          dtype: "float32",
-        },
-        {
-          name: "block_7_expand_BN/moving_variance",
-          shape: [144],
-          dtype: "float32",
-        },
-        {
-          name: "block_7_project/kernel",
-          shape: [1, 1, 144, 24],
-          dtype: "float32",
-        },
-        { name: "block_7_project_BN/gamma", shape: [24], dtype: "float32" },
-        { name: "block_7_project_BN/beta", shape: [24], dtype: "float32" },
-        {
-          name: "block_7_project_BN/moving_mean",
-          shape: [24],
-          dtype: "float32",
-        },
-        {
-          name: "block_7_project_BN/moving_variance",
-          shape: [24],
-          dtype: "float32",
-        },
-        {
-          name: "block_8_depthwise/depthwise_kernel",
-          shape: [3, 3, 144, 1],
-          dtype: "float32",
-        },
-        { name: "block_8_depthwise_BN/gamma", shape: [144], dtype: "float32" },
-        { name: "block_8_depthwise_BN/beta", shape: [144], dtype: "float32" },
-        {
-          name: "block_8_depthwise_BN/moving_mean",
-          shape: [144],
-          dtype: "float32",
-        },
-        {
-          name: "block_8_depthwise_BN/moving_variance",
-          shape: [144],
-          dtype: "float32",
-        },
-        {
-          name: "block_8_expand/kernel",
-          shape: [1, 1, 24, 144],
-          dtype: "float32",
-        },
-        { name: "block_8_expand_BN/gamma", shape: [144], dtype: "float32" },
-        { name: "block_8_expand_BN/beta", shape: [144], dtype: "float32" },
-        {
-          name: "block_8_expand_BN/moving_mean",
-          shape: [144],
-          dtype: "float32",
-        },
-        {
-          name: "block_8_expand_BN/moving_variance",
-          shape: [144],
-          dtype: "float32",
-        },
-        {
-          name: "block_8_project/kernel",
-          shape: [1, 1, 144, 24],
-          dtype: "float32",
-        },
-        { name: "block_8_project_BN/gamma", shape: [24], dtype: "float32" },
-        { name: "block_8_project_BN/beta", shape: [24], dtype: "float32" },
-        {
-          name: "block_8_project_BN/moving_mean",
-          shape: [24],
-          dtype: "float32",
-        },
-        {
-          name: "block_8_project_BN/moving_variance",
-          shape: [24],
-          dtype: "float32",
-        },
-        {
-          name: "block_9_depthwise/depthwise_kernel",
-          shape: [3, 3, 144, 1],
-          dtype: "float32",
-        },
-        { name: "block_9_depthwise_BN/gamma", shape: [144], dtype: "float32" },
-        { name: "block_9_depthwise_BN/beta", shape: [144], dtype: "float32" },
-        {
-          name: "block_9_depthwise_BN/moving_mean",
-          shape: [144],
-          dtype: "float32",
-        },
-        {
-          name: "block_9_depthwise_BN/moving_variance",
-          shape: [144],
-          dtype: "float32",
-        },
-        {
-          name: "block_9_expand/kernel",
-          shape: [1, 1, 24, 144],
-          dtype: "float32",
-        },
-        { name: "block_9_expand_BN/gamma", shape: [144], dtype: "float32" },
-        { name: "block_9_expand_BN/beta", shape: [144], dtype: "float32" },
-        {
-          name: "block_9_expand_BN/moving_mean",
-          shape: [144],
-          dtype: "float32",
-        },
-        {
-          name: "block_9_expand_BN/moving_variance",
-          shape: [144],
-          dtype: "float32",
-        },
-        {
-          name: "block_9_project/kernel",
-          shape: [1, 1, 144, 24],
-          dtype: "float32",
-        },
-        { name: "block_9_project_BN/gamma", shape: [24], dtype: "float32" },
-        { name: "block_9_project_BN/beta", shape: [24], dtype: "float32" },
-        {
-          name: "block_9_project_BN/moving_mean",
-          shape: [24],
-          dtype: "float32",
-        },
-        {
-          name: "block_9_project_BN/moving_variance",
-          shape: [24],
-          dtype: "float32",
-        },
-        { name: "bn_Conv1/gamma", shape: [16], dtype: "float32" },
-        { name: "bn_Conv1/beta", shape: [16], dtype: "float32" },
-        { name: "bn_Conv1/moving_mean", shape: [16], dtype: "float32" },
-        { name: "bn_Conv1/moving_variance", shape: [16], dtype: "float32" },
-        { name: "dense/kernel", shape: [1280, 1024], dtype: "float32" },
-        { name: "dense/bias", shape: [1024], dtype: "float32" },
-        { name: "dense_1/kernel", shape: [1024, 2], dtype: "float32" },
-        { name: "dense_1/bias", shape: [2], dtype: "float32" },
-        {
-          name: "expanded_conv_depthwise/depthwise_kernel",
-          shape: [3, 3, 16, 1],
-          dtype: "float32",
-        },
-        {
-          name: "expanded_conv_depthwise_BN/gamma",
-          shape: [16],
-          dtype: "float32",
-        },
-        {
-          name: "expanded_conv_depthwise_BN/beta",
-          shape: [16],
-          dtype: "float32",
-        },
-        {
-          name: "expanded_conv_depthwise_BN/moving_mean",
-          shape: [16],
-          dtype: "float32",
-        },
-        {
-          name: "expanded_conv_depthwise_BN/moving_variance",
-          shape: [16],
-          dtype: "float32",
-        },
-        {
-          name: "expanded_conv_project/kernel",
-          shape: [1, 1, 16, 8],
-          dtype: "float32",
-        },
-        {
-          name: "expanded_conv_project_BN/gamma",
-          shape: [8],
-          dtype: "float32",
-        },
-        { name: "expanded_conv_project_BN/beta", shape: [8], dtype: "float32" },
-        {
-          name: "expanded_conv_project_BN/moving_mean",
-          shape: [8],
-          dtype: "float32",
-        },
-        {
-          name: "expanded_conv_project_BN/moving_variance",
-          shape: [8],
-          dtype: "float32",
-        },
-      ],
-    },
-  ],
-};
